pyproject.toml
[project]
name = "relay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.7.0",
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "djoser>=2.3.1",
    "channels[daphne]>=4.2.0",
    "channels-redis>=4.2.1",
    "cuid2>=2.0.1",
    "psycopg2>=2.9.10",
    "snowflake-id>=1.0.2",
    "python-dotenv>=1.0.1",
    "resend>=2.6.0",
]
